+++ crates/client/db/src/mapping_db.rs
@@ -21,14 +21,12 @@ pub struct MappingCommitment<B: BlockT> {
 /// Allow interaction with the mapping db
 pub struct MappingDb {
     db: Arc<DB>,
-    /// Whether more information should be cached in the database.
-    cache_more_things: bool,
 }
 
 impl MappingDb {
     /// Creates a new instance of the mapping database.
-    pub(crate) fn new(db: Arc<DB>, cache_more_things: bool) -> Self {
-        Self { db, cache_more_things }
+    pub(crate) fn new(db: Arc<DB>) -> Self {
+        Self { db }
     }
 
     /// Check if the given block hash has already been processed
@@ -45,7 +43,7 @@ impl MappingDb {
     ///
     /// Under some circumstances it can return multiples blocks hashes, meaning that the result has
     /// to be checked against the actual blockchain state in order to find the good one.
-    pub fn block_hash(&self, starknet_block_hash: StarkHash) -> Result<Option<Vec<DHashT>>, DbError> {
+    pub fn substrate_block_hash(&self, starknet_block_hash: StarkHash) -> Result<Option<Vec<DHashT>>, DbError> {
         let block_mapping_col = self.db.get_column(Column::BlockMapping);
 
         match self.db.get_cf(&block_mapping_col, starknet_block_hash.encode())? {
@@ -67,12 +65,13 @@ impl MappingDb {
         let synced_mapping_col = self.db.get_column(Column::SyncedMapping);
         let block_mapping_col = self.db.get_column(Column::BlockMapping);
         let transaction_mapping_col = self.db.get_column(Column::TransactionMapping);
-        let starknet_tx_hashes_col = self.db.get_column(Column::StarknetTransactionHashesCache);
-        let starknet_block_hashes_col = self.db.get_column(Column::StarknetBlockHashesCache);
+        let starknet_tx_hashes_col = self.db.get_column(Column::StarknetTransactionHashesMapping);
+        let starknet_block_hashes_col = self.db.get_column(Column::StarknetBlockHashesMapping);
+        let starknet_block_numbers_col = self.db.get_column(Column::StarknetBlockNumberMapping);
 
         let mut transaction: WriteBatchWithTransaction<true> = Default::default();
 
-        let substrate_hashes = match self.block_hash(commitment.starknet_block_hash) {
+        let substrate_hashes = match self.substrate_block_hash(commitment.starknet_block_hash) {
             Ok(Some(mut data)) => {
                 data.push(commitment.block_hash);
                 // log::warn!(
@@ -94,19 +93,22 @@ impl MappingDb {
             transaction.put_cf(&transaction_mapping_col, &transaction_hash.encode(), &commitment.block_hash.encode());
         }
 
-        if self.cache_more_things {
-            transaction.put_cf(
-                &starknet_tx_hashes_col,
-                &commitment.starknet_block_hash.encode(),
-                &commitment.starknet_transaction_hashes.encode(),
-            );
+        transaction.put_cf(
+            &starknet_tx_hashes_col,
+            &commitment.starknet_block_hash.encode(),
+            &commitment.starknet_transaction_hashes.encode(),
+        );
 
-            transaction.put_cf(
-                &starknet_block_hashes_col,
-                &commitment.block_number.encode(),
-                &commitment.starknet_block_hash.encode(),
-            );
-        }
+        transaction.put_cf(
+            &starknet_block_hashes_col,
+            &commitment.block_number.encode(),
+            &commitment.starknet_block_hash.encode(),
+        );
+        transaction.put_cf(
+            &starknet_block_numbers_col,
+            &commitment.starknet_block_hash.encode(),
+            &commitment.block_number.encode(),
+        );
 
         self.db.write(transaction)?;
 
@@ -121,7 +123,10 @@ impl MappingDb {
     /// * `transaction_hash` - the transaction hash to search for. H256 is used here because it's a
     ///   native type of substrate, and we are sure it's SCALE encoding is optimized and will not
     ///   change.
-    pub fn block_hash_from_transaction_hash(&self, transaction_hash: StarkHash) -> Result<Option<DHashT>, DbError> {
+    pub fn substrate_block_hash_from_transaction_hash(
+        &self,
+        transaction_hash: StarkHash,
+    ) -> Result<Option<DHashT>, DbError> {
         let transaction_mapping_col = self.db.get_column(Column::TransactionMapping);
 
         match self.db.get_cf(&transaction_mapping_col, transaction_hash.encode())? {
@@ -144,16 +149,11 @@ impl MappingDb {
     ///
     /// - The cache is disabled.
     /// - The provided `starknet_hash` is not present in the cache.
-    pub fn cached_transaction_hashes_from_block_hash(
+    pub fn transaction_hashes_from_block_hash(
         &self,
         starknet_block_hash: StarkHash,
     ) -> Result<Option<Vec<StarkHash>>, DbError> {
-        let starknet_tx_hashes_col = self.db.get_column(Column::StarknetTransactionHashesCache);
-
-        if !self.cache_more_things {
-            // The cache is not enabled, no need to even touch the database.
-            return Ok(None);
-        }
+        let starknet_tx_hashes_col = self.db.get_column(Column::StarknetTransactionHashesMapping);
 
         match self.db.get_cf(&starknet_tx_hashes_col, starknet_block_hash.encode())? {
             Some(raw) => Ok(Some(Vec::<StarkHash>::decode(&mut &raw[..])?)),
@@ -161,8 +161,6 @@ impl MappingDb {
         }
     }
 
-    /// Returns the cached block hash of a given block number.
-    ///
     /// # Arguments
     ///
     /// * `block_number` - the block number to search for.
@@ -171,23 +169,34 @@ impl MappingDb {
     ///
     /// The block hash of a given block number.
     ///
-    /// This function may return `None` for two separate reasons:
-    ///
-    /// - The cache is disabled.
-    /// - The provided `starknet_hash` is not present in the cache.
-    pub fn cached_block_hash_from_block_number(
-        &self,
-        starknet_block_number: u64,
-    ) -> Result<Option<StarkHash>, DbError> {
-        let starknet_block_hashes_col = self.db.get_column(Column::StarknetBlockHashesCache);
+    /// This function may return `None` if the provided `block_number` is not present in the cache.
+    pub fn starknet_block_hash_from_block_number(&self, block_number: u64) -> Result<Option<StarkHash>, DbError> {
+        let starknet_block_hashes_col = self.db.get_column(Column::StarknetBlockHashesMapping);
 
-        if !self.cache_more_things {
-            // The cache is not enabled, no need to even touch the database.
-            return Ok(None);
-        }
-
-        match self.db.get_cf(&starknet_block_hashes_col, starknet_block_number.encode())? {
+        match self.db.get_cf(&starknet_block_hashes_col, block_number.encode())? {
             Some(raw) => Ok(Some(<StarkHash>::decode(&mut &raw[..])?)),
+            None => Ok(None),
+        }
+    }
+
+    /// # Arguments
+    ///
+    /// * `starknet_block_hash` - the block number to search for.
+    ///
+    /// # Returns
+    ///
+    /// The block hash of a given block number.
+    ///
+    /// This function may return `None` if the provided `starknet_block_hash` is not present in the
+    /// cache.
+    pub fn block_number_from_starknet_block_hash(
+        &self,
+        starknet_block_hash: StarkHash,
+    ) -> Result<Option<u64>, DbError> {
+        let starknet_block_numbers_col = self.db.get_column(Column::StarknetBlockNumberMapping);
+
+        match self.db.get_cf(&starknet_block_numbers_col, starknet_block_hash.encode())? {
+            Some(raw) => Ok(Some(u64::decode(&mut &raw[..])?)),
             None => Ok(None),
         }
     }